Validation & Preorders
Validation complete for:
- Customers
- Items
- Preorders
- Tickets
- Employees

Validating user ID and customer ID against database for tickets and
preorders still required. Currently allows any id to be entered, will
create an unhandled error if id is not valid. This will be in final
release.

Pre-orders layout is complete. Removed preorder view as all info is
available through the table. Preorders can be deleted and marked as
complete, zeroing out balance.

Tutorials WIP, will not be in beta demo.

You cannot view invoices, preorders, and tickets associated with a user
through the user page due to an Eloquent error. This will not be in for
the beta demo, but will be present in the final release.